docs: polishing work on documentation 💯
