+++ demo/components/ConceptsScreen/Screens/Overview.md
@@ -6,7 +6,7 @@
 
 <!-- Use passed routes map to include route name state in link for header -->
 
-* <Link to={props.conceptsRoutes.accessibility}>A++ Accessibility</Link> covers out of the box accessibility provided by Componentry and recommendations for creating accessible content.
+* <Link to={props.conceptsRoutes.accessibility}><Icon id="stars" /> A++ Accessibility</Link> covers out of the box accessibility provided by Componentry and recommendations for creating accessible content.
 * <Link to={props.conceptsRoutes.theming}>Theme Customization</Link> covers customizing the styles and icons for Bootstrap + Componentry using SASS variables.
 * <Link to={props.conceptsRoutes.components}>Component APIs</Link> covers the APIs Componentry components expose for composing and customizing new components.
 * <Link to={props.conceptsRoutes.architecture}>Componentry Architecture</Link> covers the design details of the library and is useful for advanced users and contributors.
