--- demo/components/universal/PageBreak/PageBreak.jsx
@@ -1,3 +1,0 @@
-import React from 'react'
-
-export default () => <div className="mb-5">&nbsp;</div>
