--- demo/components/universal/PageBreak/PageBreak.registry.js
@@ -1,2 +1,0 @@
-export const registryName = 'PageBreak'
-export { default } from './PageBreak'
