--- demo/components/universal/PropsDocs/PropsDocs.registry.js
@@ -1,2 +1,0 @@
-export const registryName = 'PropsDocs'
-export { default } from './PropsDocs'
