--- demo/components/universal/PropsDocs/PropsDocsContent/index.js
@@ -1,1 +1,0 @@
-export { default } from './PropsDocsContent'
