+++ demo/components/universal/PropsTabs/PropsTabs.registry.js
@@ -1,0 +1,2 @@
+export const registryName = 'PropsTabs'
+export { default } from './PropsTabs'
