+++ demo/components/universal/PropsTabs/TabContent/index.js
@@ -1,0 +1,1 @@
+export { default } from './TabContent'
