+++ demo/components/universal/PropsTabs/TabContent/tab-content.scss
