+++ demo/components/universal/PropsTabs/props-tabs.scss
