+++ demo/styles/application/application.scss
@@ -1,0 +1,6 @@
+/* demo/styles/application */
+
+// This icon is used with all A++ Accessible headers and is default green
+.icon.stars {
+  color: $success;
+}
