+++ demo/styles/theme.scss
@@ -4,6 +4,11 @@
 // This theme file contains the complete theme declaration for the application, it
 // can be imported into any component file.
 
+// Componentry
+
+// Use SVG sprites in demo app
+$include-icons: false;
+
 // ========================================================
 // Flatly 4.0.0 - Bootswatch
 // ========================================================
